requirements.txt
numpy~=1.22.4
emoji~=1.7.0
regex~=2022.4.24
pandas~=1.4.2
spacy~=3.3.0
wordcloud~=1.8.1
requests~=2.27.1
tweepy~=4.10.0
Django~=4.0.4
python-decouple~=3.6